Fix: arma al centro corpo + warning Enemy.cpp rotate()
1. ARMA ALZATA AL CENTRO CORPO (Player.cpp)
   PROBLEMA: l'arma appariva ai piedi del player (pos.y + 4), sembrava
   a terra invece che imbracciata.
   FIX: alzata a pos.y - 12 (centro del corpo, come se fosse
   imbracciata dal player). Lo sprite e' 64px con anchor piedi a
   (32,56), quindi il centro del corpo e' circa a pos.y - 16.
   Usiamo pos.y - 12 come altezza 'imbracciata'.
   - Movimento orizzontale: arma a destra/sinistra al centro corpo
   - Movimento giu: arma leggermente piu' bassa (pos.y + 6)
   - Movimento su: arma piu' alta (pos.y - 20)
   Applicato a entrambi i rami: sprite PNG e fallback procedurale.

2. WARNING C4244 Enemy.cpp L902/L904
   PROBLEMA: rotate(-30 + i*20) passava int a rotate(float).
   FIX: cast esplicito (float)(-30 + i*20) e (float)(30 - i*20).

+++ src/Enemy.cpp
@@ -899,9 +899,9 @@ void Enemy::renderPrimitives(sf::RenderTarget& target) const {
         sf::Color web(200, 200, 200);
         for(int i=0; i<4; i++) {
             sf::RectangleShape l1(sf::Vector2f(12.f, 2.f)); l1.setFillColor(web);
-            l1.rotate(-30 + i*20); l1.setPosition(px - 8.f, py - 2.f + (float)(i*3)); target.draw(l1);
+            l1.rotate((float)(-30 + i*20)); l1.setPosition(px - 8.f, py - 2.f + (float)(i*3)); target.draw(l1);
             sf::RectangleShape l2(sf::Vector2f(12.f, 2.f)); l2.setFillColor(web);
-            l2.rotate(30 - i*20); l2.setPosition(px + 8.f, py - 2.f + (float)(i*3)); target.draw(l2);
+            l2.rotate((float)(30 - i*20)); l2.setPosition(px + 8.f, py - 2.f + (float)(i*3)); target.draw(l2);
         }
         sf::CircleShape body(10.f); body.setFillColor(sf::Color(30, 30, 30)); body.setOutlineThickness(1.f); body.setOutlineColor(outline);
         body.setPosition(px - 10.f, py - 4.f); target.draw(body);
